first version

+++ MedSegDGSSL/metrics/accuracy.py
@@ -19,7 +19,7 @@ def compute_dice(output, target):
     output_onehot = torch.flatten(output_onehot, start_dim=2)
     
     label_onehot = torch.zeros_like(output)
-    label_onehot.scatter_(dim=1, index=target, value=1)
+    label_onehot.scatter_(dim=1, index=target.to(torch.long), value=1)
     label_onehot = torch.flatten(label_onehot, start_dim=2)
 
     dice_value = 2*(torch.sum(output_onehot*label_onehot, dim=[0, 2])+eps) / (torch.sum((label_onehot + output_onehot), dim=[0, 2]) + eps)
